prevent tabs in texi files
If global `indent-tabs-mode' is `t', then code blocks may end up
containing tabs followed by the wrong amount of spaces.  Prevent that
by advising `org-export-to-file' and org-export-to-buffer' to bind
`indent-tabs-mode' to `nil'.  (We cannot use `let' for this purpose
because this is a buffer-local value and so a let-binding may lose
its effect when changing buffers.)