Replace untabify advice
Replace the advice that was used to replace tabs in code blacks with
spaces when exporting.  It no longer (?) works and is not necessary
anymore because Org now (?) for the most part respects the value of
`indent-tabs-mode' as set in the Org buffer.

However Org still does not set `indent-tabs-mode' to nil in the buffer
used to edit a source block when appropriate.  This isn't a big deal
since it replaces the tabs when putting the edited code back into the
code block, but it's still a bit confusing.

The new kludge makes sure that `indent-tabs-mode' is set correctly.